refactor: rename seismicity to catalogs
