+++ seismostats/catalogs/rategrid.py
